[Pass][SPGO] Adding InstCombinePass and SimplifyCFGPass before SampleProfileLoaderPass when building in SPGO mode.
In Spgo mode, SampleProfileLoaderPass has to happen in the very beginning of passpipeline.
So that inlineing hasspens also very early in SampleProfileLoaderPass. At this point, If
small functions are not simplified enough, it later optimization result would be impacted
because it easily involves other opt chance in a bigger scope but actuallly the opt is
not the best choice.

Detailed dicuss is in
https://discourse.llvm.org/t/rfc-spgo-passpipelines-adding-instcombinepass-and-simplifycfgpass-before-sampleprofileloaderpass-when-building-in-spgo-mode/83340